Add `PseudoAnonymizedFaker` transformer (#523)
* Remove commented code

* Bump threshold for onehotencoder.

* Implement PSeudoAnonymizedFaker

* Finalize unit tests

* add integration tests

* Increase coverage

* Improve error messages

* Address comments

* Update docstrings

* Use uniqueness for the PseudoAnonymizer

* Remove class attribute

* Address comments and fix build

* Fix deepcopy of faker.unique

* Bump Faker version

+++ rdt/transformers/categorical.py
@@ -429,7 +429,7 @@ class LabelEncoder(BaseTransformer):
     """
 
     INPUT_SDTYPE = 'categorical'
-    OUTPUT_SDTYPES = {'value': 'integer'}
+    OUTPUT_SDTYPES = {'value': 'float'}
     DETERMINISTIC_TRANSFORM = True
     DETERMINISTIC_REVERSE = True
     COMPOSITION_IS_IDENTITY = True
